after running loop, explain my lack of space
I was able to upload all the FIA absences for the 13 species in FIA, but not in one file. The absences for the first 12 species are in the larger file in the compiled folder and the absences for the remaining 13th species is also in the folder in a distinct csv.

The loop should run without issue on the server, and can maybe be reworked to run more quickly in a function format.