give county centroids their coordinates before removing occurrences lacking coordinates
To increase our number of occurrences, I assigned county-labeled occurrences lacking coordinates with the coordinates of the county centroids. We may end up removing these points at a later time, but it may also be important to recognize the presence of a species in one of these counties, in a broad sense.

I had to play around with changing the  class of columns and removing extra numerals and text from entries, but about 3500 additional county centroid occurrences were preserved in our dataset through this step.